23.12

+++ parser.cpp
@@ -5,26 +5,26 @@ void parser::pars(char* filename)
 	std::ifstream in(filename);
 	std::string nameOfFilter;
 	while (std::getline(in, nameOfFilter)) {
-		if (nameOfFilter.find("Red") != -1) {
+		if ((int)nameOfFilter.find("Red") != -1) {
 			str.nameOfFilter = "Red";
 			nameOfFilter.erase(0, 4);
 			parsStr(nameOfFilter);
 			filterQ.push_back(str);
 		}
-		if (nameOfFilter.find("Treshold") != -1) {
+		if ((int)nameOfFilter.find("Treshold") != -1) {
 			str.nameOfFilter = "Treshold";
 			nameOfFilter.erase(0, 9);
 			parsStr(nameOfFilter);
 			filterQ.push_back(str);
 		}
-		
-		if (nameOfFilter.find("Edge") != -1) {
+
+		if ((int)nameOfFilter.find("Edge") != -1) {
 			str.nameOfFilter = "Edge";
-			nameOfFilter.erase(0, 5);	
+			nameOfFilter.erase(0, 5);
 			parsStr(nameOfFilter);
 			filterQ.push_back(str);
 		}
-		if (nameOfFilter.find("Blur") != -1) {
+		if ((int)nameOfFilter.find("Blur") != -1) {
 			str.nameOfFilter = "Blur";
 			nameOfFilter.erase(0, 5);
 			parsStr(nameOfFilter);
